Finish computeFinalDisparities
Finished the function to compute the ending optimal disparities, based on the description of such on page 42 of Felzenswalb & Huttenlocher 2006.

So far, the results are extremely poor.  There is clearly a bug somewhere in this, especially given that the whole thing tends to fade to white as the iteration count grows, rather than producing a better disparity map.

Also started work a function to compute the overall energy of the system, just to check that the energy in fact is going down after each iteration.