Complete basic version of the Nghia BP
Got this working, and in some ways it gives better results than the other version.

Kinda slow still of course.

Still not convinced about the numerical stability of this one, however.

Need to look at this more, and see if I can work out why this is so different to the other version.