Introduce Intel MKL provider
Added the MathNet Numerics Linear Algebra Intel Math Kernel NuGet package to this project.

That seems to have the effect of automatically making the appropriate dlls get published to the release folder, and then seemingly getting the program to use them automatically.

The net effect seems to be that the program runs maybe up to about 10% faster.  I suspect that the linear algebra operations being performed here are just too small and not enough of the program's total runtime for it to be of much use.

Plus, the dlls are huuuuuuuge.