Fix the stripes bug
After very helpfully being pointed to the actual issue by medusacle on the Rust Users forum, the problem that was causing the program just to output gray stripes has been fixed.

Turns out, it was just that I wasn't taking into account the y co-ordinates when computing the data costs...

Also, changed around some of the type declarations so that usizes are used by default in the places that u32s were being used previously.  This was done since I was mostly just casting from u32 to usize anyway.